+++ src/test/java/dev/orne/beans/converters/OrneBeansConvertersTest.java
@@ -1,6 +1,3 @@
-/**
- * 
- */
 package dev.orne.beans.converters;
 
 /*-
@@ -99,11 +96,46 @@ public class OrneBeansConvertersTest {
     }
 
     /**
+     * Test for {@code OrneBeansConverters#register(boolean)}.
+     */
+    @Test
+    public void testRegisterNulls() {
+        OrneBeansConverters.register(true);
+        assertNotNull(ConvertUtils.lookup(Identity.class));
+        assertNotNull(ConvertUtils.lookup(Locale.class));
+        assertNotNull(ConvertUtils.lookup(Instant.class));
+        assertNotNull(ConvertUtils.lookup(Year.class));
+        assertNotNull(ConvertUtils.lookup(YearMonth.class));
+        assertNotNull(ConvertUtils.lookup(Month.class));
+        assertNotNull(ConvertUtils.lookup(MonthDay.class));
+        assertNotNull(ConvertUtils.lookup(DayOfWeek.class));
+        assertNotNull(ConvertUtils.lookup(LocalDate.class));
+        assertNotNull(ConvertUtils.lookup(LocalTime.class));
+        assertNotNull(ConvertUtils.lookup(LocalDateTime.class));
+        assertNotNull(ConvertUtils.lookup(OffsetTime.class));
+        assertNotNull(ConvertUtils.lookup(OffsetDateTime.class));
+        assertNotNull(ConvertUtils.lookup(ZonedDateTime.class));
+        assertNotNull(ConvertUtils.lookup(ZoneId.class));
+        assertNotNull(ConvertUtils.lookup(ZoneOffset.class));
+        assertNotNull(ConvertUtils.lookup(Duration.class));
+        assertNotNull(ConvertUtils.lookup(Period.class));
+    }
+
+    /**
      * Test for {@code OrneBeansConverters#registerBeansConversors()}.
      */
     @Test
     public void testRegisterBeans() {
         OrneBeansConverters.registerBeansConversors();
+        assertNotNull(ConvertUtils.lookup(Identity.class));
+    }
+
+    /**
+     * Test for {@code OrneBeansConverters#registerBeansConversors(boolean)}.
+     */
+    @Test
+    public void testRegisterBeansNulls() {
+        OrneBeansConverters.registerBeansConversors(true);
         assertNotNull(ConvertUtils.lookup(Identity.class));
     }
 
@@ -117,11 +149,44 @@ public class OrneBeansConvertersTest {
     }
 
     /**
+     * Test for {@code OrneBeansConverters#registerUtilConversors(boolean)}.
+     */
+    @Test
+    public void testRegisterUtilNulls() {
+        OrneBeansConverters.registerUtilConversors(true);
+        assertNotNull(ConvertUtils.lookup(Locale.class));
+    }
+
+    /**
      * Test for {@code OrneBeansConverters#registerTimeConversors()}.
      */
     @Test
     public void testRegisterTime() {
         OrneBeansConverters.registerTimeConversors();
+        assertNotNull(ConvertUtils.lookup(Instant.class));
+        assertNotNull(ConvertUtils.lookup(Year.class));
+        assertNotNull(ConvertUtils.lookup(YearMonth.class));
+        assertNotNull(ConvertUtils.lookup(Month.class));
+        assertNotNull(ConvertUtils.lookup(MonthDay.class));
+        assertNotNull(ConvertUtils.lookup(DayOfWeek.class));
+        assertNotNull(ConvertUtils.lookup(LocalDate.class));
+        assertNotNull(ConvertUtils.lookup(LocalTime.class));
+        assertNotNull(ConvertUtils.lookup(LocalDateTime.class));
+        assertNotNull(ConvertUtils.lookup(OffsetTime.class));
+        assertNotNull(ConvertUtils.lookup(OffsetDateTime.class));
+        assertNotNull(ConvertUtils.lookup(ZonedDateTime.class));
+        assertNotNull(ConvertUtils.lookup(ZoneId.class));
+        assertNotNull(ConvertUtils.lookup(ZoneOffset.class));
+        assertNotNull(ConvertUtils.lookup(Duration.class));
+        assertNotNull(ConvertUtils.lookup(Period.class));
+    }
+
+    /**
+     * Test for {@code OrneBeansConverters#registerTimeConversors(boolean)}.
+     */
+    @Test
+    public void testRegisterTimeNulls() {
+        OrneBeansConverters.registerTimeConversors(true);
         assertNotNull(ConvertUtils.lookup(Instant.class));
         assertNotNull(ConvertUtils.lookup(Year.class));
         assertNotNull(ConvertUtils.lookup(YearMonth.class));
@@ -168,12 +233,49 @@ public class OrneBeansConvertersTest {
     }
 
     /**
+     * Test for {@code OrneBeansConverters#register(ConvertUtilsBean, boolean)}.
+     */
+    @Test
+    public void testRegisterInBeanNulls() {
+        final ConvertUtilsBean converter = new ConvertUtilsBean();
+        OrneBeansConverters.register(converter, true);
+        assertNotNull(converter.lookup(Identity.class));
+        assertNotNull(converter.lookup(Locale.class));
+        assertNotNull(converter.lookup(Instant.class));
+        assertNotNull(converter.lookup(Year.class));
+        assertNotNull(converter.lookup(YearMonth.class));
+        assertNotNull(converter.lookup(Month.class));
+        assertNotNull(converter.lookup(MonthDay.class));
+        assertNotNull(converter.lookup(DayOfWeek.class));
+        assertNotNull(converter.lookup(LocalDate.class));
+        assertNotNull(converter.lookup(LocalTime.class));
+        assertNotNull(converter.lookup(LocalDateTime.class));
+        assertNotNull(converter.lookup(OffsetTime.class));
+        assertNotNull(converter.lookup(OffsetDateTime.class));
+        assertNotNull(converter.lookup(ZonedDateTime.class));
+        assertNotNull(converter.lookup(ZoneId.class));
+        assertNotNull(converter.lookup(ZoneOffset.class));
+        assertNotNull(converter.lookup(Duration.class));
+        assertNotNull(converter.lookup(Period.class));
+    }
+
+    /**
      * Test for {@code OrneBeansConverters#registerBeansConversors(ConvertUtilsBean)}.
      */
     @Test
     public void testRegisterBeansInBean() {
         final ConvertUtilsBean converter = new ConvertUtilsBean();
         OrneBeansConverters.registerBeansConversors(converter);
+        assertNotNull(converter.lookup(Identity.class));
+    }
+
+    /**
+     * Test for {@code OrneBeansConverters#registerBeansConversors(ConvertUtilsBean, boolean)}.
+     */
+    @Test
+    public void testRegisterBeansInBeanNulls() {
+        final ConvertUtilsBean converter = new ConvertUtilsBean();
+        OrneBeansConverters.registerBeansConversors(converter, true);
         assertNotNull(converter.lookup(Identity.class));
     }
 
@@ -188,12 +290,47 @@ public class OrneBeansConvertersTest {
     }
 
     /**
+     * Test for {@code OrneBeansConverters#registerUtilConversors(ConvertUtilsBean, boolean)}.
+     */
+    @Test
+    public void testRegisterUtilInBeanNulls() {
+        final ConvertUtilsBean converter = new ConvertUtilsBean();
+        OrneBeansConverters.registerUtilConversors(converter, true);
+        assertNotNull(converter.lookup(Locale.class));
+    }
+
+    /**
      * Test for {@code OrneBeansConverters#registerTimeConversors(ConvertUtilsBean)}.
      */
     @Test
     public void testRegisterTimeInBean() {
         final ConvertUtilsBean converter = new ConvertUtilsBean();
         OrneBeansConverters.registerTimeConversors(converter);
+        assertNotNull(converter.lookup(Instant.class));
+        assertNotNull(converter.lookup(Year.class));
+        assertNotNull(converter.lookup(YearMonth.class));
+        assertNotNull(converter.lookup(Month.class));
+        assertNotNull(converter.lookup(MonthDay.class));
+        assertNotNull(converter.lookup(DayOfWeek.class));
+        assertNotNull(converter.lookup(LocalDate.class));
+        assertNotNull(converter.lookup(LocalTime.class));
+        assertNotNull(converter.lookup(LocalDateTime.class));
+        assertNotNull(converter.lookup(OffsetTime.class));
+        assertNotNull(converter.lookup(OffsetDateTime.class));
+        assertNotNull(converter.lookup(ZonedDateTime.class));
+        assertNotNull(converter.lookup(ZoneId.class));
+        assertNotNull(converter.lookup(ZoneOffset.class));
+        assertNotNull(converter.lookup(Duration.class));
+        assertNotNull(converter.lookup(Period.class));
+    }
+
+    /**
+     * Test for {@code OrneBeansConverters#registerTimeConversors(ConvertUtilsBean, boolean)}.
+     */
+    @Test
+    public void testRegisterTimeInBeanNullst() {
+        final ConvertUtilsBean converter = new ConvertUtilsBean();
+        OrneBeansConverters.registerTimeConversors(converter, true);
         assertNotNull(converter.lookup(Instant.class));
         assertNotNull(converter.lookup(Year.class));
         assertNotNull(converter.lookup(YearMonth.class));